finish moving sync code out of @cocalc/project
- this breaks compute, will fix later
