+++ src/packages/project/browser-websocket/api.ts
@@ -21,8 +21,8 @@ import { nbconvert as jupyter_nbconvert } from "../jupyter/convert";
 import { lean, lean_channel } from "../lean/server";
 import { jupyter_strip_notebook } from "@cocalc/jupyter/nbgrader/jupyter-parse";
 import { jupyter_run_notebook } from "@cocalc/jupyter/nbgrader/jupyter-run";
-import { synctable_channel } from "../sync/server";
-import { callSyncDoc } from "@cocalc/sync-server/syncdocs-manager";
+import { synctable_channel } from "@cocalc/sync-server/server/server";
+import { callSyncDoc } from "@cocalc/sync-server/server/syncdocs-manager";
 import { terminal } from "@cocalc/terminal";
 import { x11_channel } from "@cocalc/sync-server/x11";
 import { canonical_paths } from "./canonical-path";
