+++ src/packages/sync-server/server/compute-server-open-file-tracking.ts
@@ -4,7 +4,7 @@ Manage the state of open files in the compute servers syncdb sync'd file.
 NOTE: terminals aren't handled at all here, since they don't have a syncdoc.
 */
 
-import type { SyncDocs } from "@cocalc/sync-server/syncdocs-manager";
+import type { SyncDocs } from "@cocalc/sync-server/server/syncdocs-manager";
 import type { SyncDB } from "@cocalc/sync/editor/db/sync";
 import { once } from "@cocalc/util/async-utils";
 import { meta_file, auxFileToOriginal } from "@cocalc/util/misc";
