+++ src/packages/sync-server/server/open-synctables.ts
